AV-55910: Create an MCP client for Galley
This PR creates a MCP client for the k8s controller that talks to
galley via a insecure grpc channel.

Note: This is a initial PR and improvments can be expected over this.
Some of the improvements could be w.r.t using logging instead of
fmt.

If the istio services are running with mTLS, we need to have the
ability to create certs and authenticate to galley.

Presently supported objects:

 - ServiceEntry
 - VirtualService
 - Gateway