AV-56761: Re-factor of go controller phase 1
This is a phased re-factoring effort of the present k8s go controller
code. This phase creates a utils package and moves some of the go
files to this package. Due to this movement, there's a huge re-factor
that was needed in the other files. One significant change is the log
file init.