AV-59563: Some ehancments to the mcp client (#79)
* AV-59563: Some ehancments to the mcp client

This patch does the following:
 - Defines some native data structures to start parsing the mcp events for CRDs.
 - Comes up with a mechanism to detect object updates/creates/deletes for
   non-incremental MCP client.

Next steps:
 - Define a queue where each of the objects are pushed for processing by
workers

* incorporated code-review comments

* AV-59564: Implement a rate limiting queue to enqueue objects

This patch introduces a rate limiting queue by utilizing the kube
workqueue. We also create a operation map to always pick the latest
operation to perform on an object. Note, the mcp client is not blocked
even if the AVI APIs are working relatively slower. It pushes the object
into a workqueue, the workqueue is then processed by another go routine.