AV-62490, AV-62640: Cleaning up the MCP client
This PR proposes fixes around the MCP client. We use the new
data model to calculate updates and do things in a more behavorial
approach making the code cleaner and easily maintainable.

Also introduced are tests that would validate the schema methods.

Next job would be to connect the MCP updates with the kubernetes
updates using a common namespace hashed work queue.