AV-62490: Common inteface to queue events
This commit introduces a common queuing interface to publish
events from both MCP and kubernetes informers. The hashing mechanism
of the kubernetes keys will be upgraded to namespace based hashes.

Some test cases will be added to try out the end to end scenarios.

The Queue logic is done by keeping in mind that there would be
another few queues that would be built over time, that would be
used across various Layers.