Cleaning up the MCP client + Introduction of common queuing interface. (#84)
* AV-62490, AV-62640: Cleaning up the MCP client

This PR proposes fixes around the MCP client. We use the new
data model to calculate updates and do things in a more behavorial
approach making the code cleaner and easily maintainable.

Also introduced are tests that would validate the schema methods.

Next job would be to connect the MCP updates with the kubernetes
updates using a common namespace hashed work queue.

* AV-62490: Common inteface to queue events

This commit introduces a common queuing interface to publish
events from both MCP and kubernetes informers. The hashing mechanism
of the kubernetes keys will be upgraded to namespace based hashes.

Some test cases will be added to try out the end to end scenarios.

The Queue logic is done by keeping in mind that there would be
another few queues that would be built over time, that would be
used across various Layers.

* Cleanups and unit tests

This commit gets rid of some of the dead wood.
Adds unit tests for the MCP client, and tests the overall Ingestion
Layer workflow.

Added are some gateway related tests as well.

* Added more tests

Added tests to do parallel updates on the store.
Also added tests on queue draining.