Fix import cycle error/code duplication
This commit addresses the issues with import cycle errors that were
caused due to the queue implementation. As a consequence to that the
queue had to be replicated in each layer. With this present model,
the queue package resides in the utils package and is used across
layers.

This also fixes some code review comments around package naming
conventions.